requirements.txt
# Python dependencies for Firewall GitOps validation scripts
pyyaml>=6.0
jsonschema>=4.0
attrs>=22.0
